Extract subprocess systemd lifecycle into its own module
